Added support for nested aggregates where the last element of the nested aggregate is a flat array.  For example, we can now deal with:
truct Ptr<`a> {
  <`i::I>               // there exists some unsigned integer i
  tag_t<`i> size;       // size == i
  `a elts[valueof(`i)]; // numelts(elts) == i
};
struct OfsPtr<`a> {
  int ofs;
  struct Ptr<`a> ptr;
};
typedef struct OfsPtr<`a>@`r ofsptr_t<`a,`r>;

And we can create arbitrarily-sized ofspointers, as in

ofsptr_t<int> ofsvectorn(tag_t n, int x) {
  return new OfsPtr{.ofs = 0, .ptr={.size = n, .elts = {for i < n : x}}};
}

To support this, I had to modify the code generator to
support passing back the bound on the comprehension in the
nested aggregate to the outer one.  I also had to allow the outer
aggregate to pass in the lhs to assign to when initializing the
nested comprehension.  This revealed a limitation to my last
fix to the boxing stuff in tovc, which I've now fixed.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2653 581db07f-6e62-4d40-9733-d432c4eebb07