Fixed the pattern matching stuff to be smarter about deep copying so that relations stuff is propagated correctly to code generation. In essence, if I have
let x = e

the corresponding "access expression" for x is e itself.  Rather than
deep copy this e, we share it, and thus relations will be set properly.
If I have any sort of nested pattern, then I deep copy it.  So, for

let x as $(y,z) = p;

The access expression for x is p (shared), while for y and z they
are p[0] and p[1] (deep copied).

All the changes to the genfiles are the result of eliminating spurious
null checks that were introduced in the last check in.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2671 581db07f-6e62-4d40-9733-d432c4eebb07