Added support for unconsuming access expressions of pattern-matched variables at the end of the pattern scope.  For example:
struct foo { int *`U x; int y; };

void f(struct foo *`U f) {
  { let &foo{.x=x,...} = f; // consumes f->x
    printf("%d\n",*x);
  } // unconsumes f->x
  Core::ufree(f->x);
  Core::ufree(f);
}

This works similarly to the noconsume(x) attribute.  When initializing a
pattern variable, we do two things:

1) We set its rval to be the same as that of its access expression.  In the
   above example, we thus initialize rval(x) to be rval(f->x).  In the case
   that the access expression is a unique pointer it will get consumed (so
   in the example rval(f->x) is now consumed).

2) We give the rval a "name" in the flow.

At the end of the pattern scope, we check that each unique pointer pattern
variable's rval still has the same name and has not been consumed, and if
so, we unconsume the access expression's rval, and then join it to the
pattern variable's current rval.  In the example we haven't changed x's
rval, so it still has its original name and hasn't been consumed, and thus
we restore f->x's rval to its original value joined with the current rval of
x.

The join is critical in the case that we've expanded out unique locations
within the pattern block and then freed them.  This will end up changing the
initialization level of the access expression so that its subpointers can't
get freed.  Also note that if we didn't do the join (e.g. we could just
restore it to UnknownR(initlevel(rval(pat-variable)))), we'd lose useful
pointer sharing information that existed before the match.

Knowing when to check for unconsuming is tricky for unstructured control
flow.  To handle this, we add to the environment a list of pattern
variables, their access expressions, and the statement in which the
variables are bound.  Then, at each break, goto, continue, or fallthru, we
use the statement encloser information to check if the flow exits the
current scope, and if so we attempt an unconsume.  This is done by making
sure that the transitive closure of the enclosers of the
destination of break/goto/etc. never includes the block in which the
variables are bound.

There were also some tricks to properly deal with unconsuming expressions
that are nested.  Basically I needed to add a flag to anal_Lexp to allow one
to dereference through a consumed pointer to find a place.

I fixed a bug with propagating named locations, and fixed the tcutil bug
that Bart found.  I also eliminated anal_stmt_exp so as to reduce duplicated
code.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2674 581db07f-6e62-4d40-9733-d432c4eebb07