Fixed bug with region h = open(key).  The current code in the flow analysis didn't deal with unconsuming quite right, and further didn't support unconsuming when you exited the block in an unstructured way, like via a goto.  I fixed this by refactoring the unconsume stuff for pattern variables, and then using the same infrastructure to unconsume opened keys.
Fixing this bug revealed a bug in the garbage collector for
tests/scheme.cyc.  Essentially it was creating the new state too
early.  If this fix breaks the cyclone_internal/scheme, you can
look at the fixed tests one as a guide.

I also took a shot a docs for dynregion, and added a dynregion test
case to tests.  I still need to finish up docs for unique pointers
and such before the release.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2856 581db07f-6e62-4d40-9733-d432c4eebb07