Changes to make it possible to extern "C include" more linux kernel headers:
When parsing extern "C" { ... } or extern "C include" { ... }, we
allow Cyclone keywords (e.g., new, malloc, etc.) to be treated as
normal identifiers.  Note that this means that if you use a Cyclone
keyword in the type of a C function, then it won't parse right.  You
can work around this by doing a typedef outside of the "C"/"C include"
(see for example include/core.h).

The aligned attribute now takes an optional expression (instead of an
integer) since GCC seems to support this.  I don't bother to check
that it makes sense.  GCC will catch it if it doesn't.

__const__ is now treated as the keyword const.

The asm lexing hack now admits "asm __volatile__ (...)" as well as
"asm volatile (...)"

It's no longer an error for extern "C include" variables to be
declared extern but have an initializer.

Fixed bug that required extern "C include" things to be initialized.

Finally fixed the parser so that we deal with unsigned, and (unsigned)
long long constants properly.  (Well, at least I believe that I'm
dealing with things properly.)  We still need to modify evexp.cyc
to use long long's and we also need to get rid of any "int = long int =
32-bits" and "long long = 64 bits" assumptions in the compiler.
But I'll leave that for another day...

Added a function to Absynpp to convert unsigned long long's to strings
so that we can pretty print or dump them.  We still need to
modify the printf and scanf functions to deal with these
guys (e.g., %lld).


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2912 581db07f-6e62-4d40-9733-d432c4eebb07