Changed how let-patterns are compiled to avoid generating labels or goto's.  The reason is that we insert let-alias declarations around function calls within the context of a statement expression as part of the alias inference.  However, the tovc code tries to eliminate this statement expression and in the process, manages to split the labels and gotos into separate functions.
So, the real problem is with tovc.  I would prefer that we
*not* use tovc by default.  I believe that GCC is actually
working correctly w.r.t. statement expressions, but that we
have to be careful when pretty-printing/dumping the code
to ensure that statement expressions actually end syntactically
with an expression.

I would like to avoid tovc for two reasons:  First, it's broken
in general because if it encounters a goto/break/continue/return
within the statement expression, we'll get the wrong code.
Second, it makes simple things, like the inserted alias code,
relatively expensive.

I also fixed a "bug" in new_control_flow.cyc due to a cascading
error.  In some situations, we can get a Dict::Absent exception
when an error has occurred previously.  We now catch that
exception, verify that an error has occurred, and then move
on to checking the next top-level declaration.  If an error
hasn't occurred, we re-throw the exception.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2968 581db07f-6e62-4d40-9733-d432c4eebb07