Modified how we implement e1 :=: e2.  Rather than emit code for each subexpression, take the address of each result, and call an external swap function, we emit the code inline.  To keep things simple, I replace e1 :=: e2 with a Cyclone expression
({ typeof(e1) tmp1 = e1;
   typeof(e2) tmp2 = e2;
   e1 = tmp2; e2 = tmp1; })

This assumes that e1 and e2 are lvalues, which the typechecker
already ensures.  Then I just run the code generator on this
expression.  This takes care of the various boundary cases
nicely, e.g., when e1 or e2 is a dereference of a tagged union.
I need the extra temporary to deal with the case that there's a
match failure in either e1 or e2 for tagged unions, so that the
swap doesn't happen partially.

The only snafu is that I have to deep copy e1 and e2 for the
latter two assignments, since code generation is different for
lhs and rhs expressions of tagged unions.  The deep copy must
preserve the relations so that bounds-check and null-check
elimination still works.  I added a function to relations that is
currently useless, but will probably come in handy when we remove
annotations and go to hashtables.

I added various tests to check all this new behavior.

Also made a small change to rstrdup to work with TR-kinded
regions.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3034 581db07f-6e62-4d40-9733-d432c4eebb07