re-run ./configure (there's a new compiler file)
This is a major check-in that largely moves the translation of aggregates and
comprehensions to a new pass (remove_aggregates.cyc) between toc.cyc
and toseqc.cyc.  There are a couple lingering things to clean up as
described below, but the tests pass and everything bootstraps.

The basic idea is that toc.cyc just leaves Aggregate_e,
Comprehension_e, Array_e, and ComprehensionNoinit_e alone (of course
it translates their components) except for doing the type translation,
generating struct-hack types, etc.  Then remove_aggregates creates the
necessary code for initializing a struct/array.  There are fewer cases
because datatypes, tuples, and anonymous structs have all been
translated to Aggregate_e by this point.

For New_e, we annotate the expression with the destination that
toc.cyc generated.  That way, toc.cyc can create the malloc-code, but
remove_aggregates can initialize the heap memory.

This change makes the compiler code base a little smaller and
substantially cleans up toc.cyc.  Alas, toc.cyc is still a mess
with respect to generating types for struct-hacks, etc.  I'm still not
clear when we should be aligning the last field and when not.

As a small change, the output now never has anonymous struct types.

Lingering things that may annoy you:

* I need to insert a couple more casts to eliminate warnings from
GCC.  So there are new warnings for tests/structarray.cyc.  I think they're
(a) innocuous and (b) fairly easy to get rid of.

* tests/scheme.cyc did not compile until I changed
initial_static_env.  Basically, a kludge from toc.cyc I removed was
hiding the fact that a cast did not have the intended effect in
guiding the type-checker.

* I'm sure I introduced bugs in the interest of getting the compiler
into a better state.  (Gotta break eggs to make an omelet.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3103 581db07f-6e62-4d40-9733-d432c4eebb07