* Changed malloc to allow allocating arrays that involve (nullable) pointers.  We simply convert the malloc to a calloc.
The next few changes are aimed at providing an alternative language
(code named Tempest) that uses fat pointers by default, but that
interoperates with Cyclone and has a lower cost in terms of porting
time.  

* Added new pragmas __tempest_on__ and __tempest_off__.  Will probably
change this to a scoped sort of declaration so that we can wrap
header files with the appropriate thing.

* Changed parser so that when in a tempest section, we parse a vanilla
pointer (*) as if it were a fat pointer.  You can still use
*@thin or @ to get a thin pointer.  

* Added new flags to the environment so we can tell when we're processing
tempest code.

* We now support a limited form of region inference for tempest functions.
Basically, when processing a tempest function, we do not generalize the
evars from the prototype (nor do we force the result to be in the heap
region.)  During type-checking of the function, these evars may become
constrained to be equal to one another (e.g., if you return an argument.)
After type-checking the function, we then generalize the evars.  

This happens to work even if you have an earlier prototype for the
function (but for very, very obscure reasons -- basically, we do not
generalize the prototype.  This is a bug that needs to be fixed and
thought about some more.)  It will not work if there's a cycle between
two functions, and similar to HM-inference, it will not handle
polymorphic recursion.

We might consider allowing the inference for Cyclone static functions...

* With these few changes, you can wrap tests/shootout/matrix.gcc with
a __tempest_on__; (after the standard #includes) and it compiles with
no further changes.  Of course, the code you get is lousy because it's
all fat pointers, and the error messages are very confusing when
things do go wrong.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3403 581db07f-6e62-4d40-9733-d432c4eebb07