Fixed bug in callgraph code.  The graph library's scc function doesn't like it when a node is a target but hasn't been entered into the dictionary.  This arose because extern "C include" functions aren't entered.  The fix for now is to not enter these into the callgraph (which I'm consructing for type inference purposes anyway.)
Fixed bug in error reporting so that we don't get obscure @true/@false
stuff when zeroterm or nullable qualifiers on pointers don't match.
I think the right thing to do here is to rewrite unify_it so it 
returns an option, with NULL being the success case and otherwise,
we return a struct containing the relevant info.

A few hacks on the relations stuff, not too important for now.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3430 581db07f-6e62-4d40-9733-d432c4eebb07