Changed the runtime and cyc_include.h to use 'fat' instead of 'dyneither'.  That's the hard part for bootstrappoing.  The easy part -- yet to be done -- is changing various variable names in the compiler in an analogous way.
Mike, it appears that after "make clean_nogc", the ensuing "make all"
is running a configure script and building a collector.  Is there a
simple Makefile fix to avoid this?

Changed Makefile.in.in so that relations and pratt are no longer used
for cycdoc and buildlib.  They were getting built in src instead of
build/boot.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3451 581db07f-6e62-4d40-9733-d432c4eebb07