Moved cifc (checking the cyclone_override block of extern "C include") phase into typechecker. Previously, the override checking was being done in the binding phase. I've moved it to the type checking phase instead. The reason is basically typedef expansion. For instance, suppose you have the following:
extern "C include" {
   int foo(unsigned int a);
}
cyclone_override {
   int foo(sizeof_t<`a> a);
}

This should work, because sizeof_t<`a> is represented as an unsigned
int. But, the previous override checking only ensured that the C
representation of the types in the override block were unifiable with
that in the extern "C include" section. In particular, there was no
typedef expansion being done to check representation consistency --
which is what is needed here.

The other issue is the following -- suppose you have this simple,
though pointless, example

extern "C include" {
   typedef unsigned int cyc_unsigned;
   int foobar(unsigned int a);
}
cyclone_override {
   int foobar(cyc_unsigned a);
}

Here, to ensure that the override is representation consistent with
the original declaration we need to expand the typedef for
cyc_unsigned. But this means that the original decls in the extern "C
include" block have to be typechecked before the types in the override
block are expanded. Once these override types are expanded and checked
for consistency, we need to type check the original declarations (this
time imperatively updated with the overrides) once again.

Of course, this kind of implies that we should repeat this process
until we reach a fixed point -- but I thought that was overkill. It's
unlikely that such kinds of overrides would actually be needed. So, I
do only one iteration and then use the merged overridden decls as
before.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3475 581db07f-6e62-4d40-9733-d432c4eebb07