* moved and re-named the function that checks whether an expression is side-effect free to Absyn::no_side_effects_exp().
* added an @assert(e) expression where e is a boolean, side-effect-free
expression.  Don't use it yet.  The idea is that this is a static
assertion that you have to have the flow analysis validate.  It
compiles down to 1 (assuming it gets validated.)  Right now, if
you turn on --vcgen then it will attempt to validate the @asserts,
but I have not yet integrated the constraint solver, so you only
get relativley syntactic, resolution-based results.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3491 581db07f-6e62-4d40-9733-d432c4eebb07